refactor: merge RateLimitPresenter into LLMProviderPresenter
