+++ src/main/presenter/llmProviderPresenter/index.ts
@@ -26,7 +26,7 @@ import { OllamaProvider } from './providers/ollamaProvider'
 import { AnthropicProvider } from './providers/anthropicProvider'
 import { DoubaoProvider } from './providers/doubaoProvider'
 import { ShowResponse } from 'ollama'
-import { CONFIG_EVENTS } from '@/events'
+import { CONFIG_EVENTS, RATE_LIMIT_EVENTS } from '@/events'
 import { TogetherProvider } from './providers/togetherProvider'
 import { GrokProvider } from './providers/grokProvider'
 import { GroqProvider } from './providers/groqProvider'
@@ -38,7 +38,29 @@ import { OpenRouterProvider } from './providers/openRouterProvider'
 import { MinimaxProvider } from './providers/minimaxProvider'
 import { AihubmixProvider } from './providers/aihubmixProvider'
 import { _302AIProvider } from './providers/_302AIProvider'
-import { RateLimitPresenter } from '../rateLimitPresenter'
+
+// 速率限制配置接口
+interface RateLimitConfig {
+  qpsLimit: number
+  enabled: boolean
+}
+
+// 队列项接口
+interface QueueItem {
+  id: string
+  timestamp: number
+  resolve: () => void
+  reject: (error: Error) => void
+}
+
+// 提供商速率限制状态接口
+interface ProviderRateLimitState {
+  config: RateLimitConfig
+  queue: QueueItem[]
+  lastRequestTime: number
+  isProcessing: boolean
+}
+
 // 流的状态
 interface StreamState {
   isGenerating: boolean
@@ -64,11 +86,17 @@ export class LLMProviderPresenter implements ILlmProviderPresenter {
     maxConcurrentStreams: 10
   }
   private configPresenter: ConfigPresenter
-  private rateLimitPresenter: RateLimitPresenter
+
+  // 速率限制相关属性
+  private providerRateLimitStates: Map<string, ProviderRateLimitState> = new Map()
+  private readonly DEFAULT_RATE_LIMIT_CONFIG: RateLimitConfig = {
+    qpsLimit: 0.1,
+    enabled: false
+  }
 
   constructor(configPresenter: ConfigPresenter) {
     this.configPresenter = configPresenter
-    this.rateLimitPresenter = new RateLimitPresenter(configPresenter)
+    this.initializeProviderRateLimitConfigs()
     this.init()
     // 监听代理更新事件
     eventBus.on(CONFIG_EVENTS.PROXY_RESOLVED, () => {
@@ -77,6 +105,21 @@ export class LLMProviderPresenter implements ILlmProviderPresenter {
         provider.onProxyResolved()
       }
     })
+  }
+
+  private initializeProviderRateLimitConfigs(): void {
+    const providers = this.configPresenter.getProviders()
+    for (const provider of providers) {
+      if (provider.rateLimit) {
+        this.setProviderRateLimitConfig(provider.id, {
+          enabled: provider.rateLimit.enabled,
+          qpsLimit: provider.rateLimit.qpsLimit
+        })
+      }
+    }
+    console.log(
+      `[LLMProviderPresenter] Initialized rate limit configs for ${providers.length} providers`
+    )
   }
 
   private init() {
@@ -212,7 +255,7 @@ export class LLMProviderPresenter implements ILlmProviderPresenter {
     const enabledProviders = Array.from(this.providers.values()).filter(
       (provider) => provider.enable
     )
-    this.rateLimitPresenter.onProvidersUpdated(providers)
+    this.onProvidersUpdated(providers)
 
     // Initialize provider instances sequentially to avoid race conditions
     for (const provider of enabledProviders) {
@@ -281,10 +324,33 @@ export class LLMProviderPresenter implements ILlmProviderPresenter {
    * 更新 provider 的速率限制配置
    */
   updateProviderRateLimit(providerId: string, enabled: boolean, qpsLimit: number): void {
-    this.rateLimitPresenter.updateProviderConfig(providerId, {
-      enabled,
-      qpsLimit
-    })
+    let finalConfig = { enabled, qpsLimit }
+    if (
+      finalConfig.qpsLimit !== undefined &&
+      (finalConfig.qpsLimit <= 0 || !isFinite(finalConfig.qpsLimit))
+    ) {
+      if (finalConfig.enabled === true) {
+        console.warn(
+          `[LLMProviderPresenter] Invalid qpsLimit (${finalConfig.qpsLimit}) for provider ${providerId}, disabling rate limit`
+        )
+        finalConfig.enabled = false
+      }
+      const provider = this.configPresenter.getProviderById(providerId)
+      finalConfig.qpsLimit = provider?.rateLimit?.qpsLimit ?? 0.1
+    }
+    this.setProviderRateLimitConfig(providerId, finalConfig)
+    const provider = this.configPresenter.getProviderById(providerId)
+    if (provider) {
+      const updatedProvider: LLM_PROVIDER = {
+        ...provider,
+        rateLimit: {
+          enabled: finalConfig.enabled,
+          qpsLimit: finalConfig.qpsLimit
+        }
+      }
+      this.configPresenter.setProviderById(providerId, updatedProvider)
+      console.log(`[LLMProviderPresenter] Updated persistent config for ${providerId}`)
+    }
   }
 
   /**
@@ -296,10 +362,10 @@ export class LLMProviderPresenter implements ILlmProviderPresenter {
     queueLength: number
     lastRequestTime: number
   } {
-    const config = this.rateLimitPresenter.getProviderConfig(providerId)
-    const currentQps = this.rateLimitPresenter.getCurrentQps(providerId)
-    const queueLength = this.rateLimitPresenter.getQueueLength(providerId)
-    const lastRequestTime = this.rateLimitPresenter.getLastRequestTime(providerId)
+    const config = this.getProviderRateLimitConfig(providerId)
+    const currentQps = this.getCurrentQps(providerId)
+    const queueLength = this.getQueueLength(providerId)
+    const lastRequestTime = this.getLastRequestTime(providerId)
 
     return {
       config,
@@ -321,7 +387,16 @@ export class LLMProviderPresenter implements ILlmProviderPresenter {
       lastRequestTime: number
     }
   > {
-    return this.rateLimitPresenter.getAllProviderStatus()
+    const status: Record<string, any> = {}
+    for (const [providerId, state] of this.providerRateLimitStates) {
+      status[providerId] = {
+        config: state.config,
+        currentQps: this.getCurrentQps(providerId),
+        queueLength: state.queue.length,
+        lastRequestTime: state.lastRequestTime
+      }
+    }
+    return status
   }
 
   isGenerating(eventId: string): boolean {
@@ -438,11 +513,11 @@ export class LLMProviderPresenter implements ILlmProviderPresenter {
         try {
           console.log(`[Agent Loop] Iteration ${toolCallCount + 1} for event: ${eventId}`)
           const mcpTools = await presenter.mcpPresenter.getAllToolDefinitions(enabledMcpTools)
-          const canExecute = this.rateLimitPresenter.canExecuteImmediately(providerId)
+          const canExecute = this.canExecuteImmediately(providerId)
           if (!canExecute) {
-            const config = this.rateLimitPresenter.getProviderConfig(providerId)
-            const currentQps = this.rateLimitPresenter.getCurrentQps(providerId)
-            const queueLength = this.rateLimitPresenter.getQueueLength(providerId)
+            const config = this.getProviderRateLimitConfig(providerId)
+            const currentQps = this.getCurrentQps(providerId)
+            const queueLength = this.getQueueLength(providerId)
 
             yield {
               type: 'response',
@@ -458,9 +533,9 @@ export class LLMProviderPresenter implements ILlmProviderPresenter {
               }
             }
 
-            await this.rateLimitPresenter.executeWithRateLimit(providerId)
+            await this.executeWithRateLimit(providerId)
           } else {
-            await this.rateLimitPresenter.executeWithRateLimit(providerId)
+            await this.executeWithRateLimit(providerId)
           }
 
           // Call the provider's core stream method, expecting LLMCoreStreamEvent
@@ -1332,6 +1407,195 @@ export class LLMProviderPresenter implements ILlmProviderPresenter {
           normalized: false
         },
         errorMsg: error instanceof Error ? error.message : String(error)
+      }
+    }
+  }
+
+  private setProviderRateLimitConfig(providerId: string, config: Partial<RateLimitConfig>): void {
+    const currentState = this.providerRateLimitStates.get(providerId)
+    const newConfig = {
+      ...this.DEFAULT_RATE_LIMIT_CONFIG,
+      ...currentState?.config,
+      ...config
+    }
+    if (!currentState) {
+      this.providerRateLimitStates.set(providerId, {
+        config: newConfig,
+        queue: [],
+        lastRequestTime: 0,
+        isProcessing: false
+      })
+    } else {
+      currentState.config = newConfig
+    }
+    console.log(`[LLMProviderPresenter] Updated rate limit config for ${providerId}:`, newConfig)
+    eventBus.send(RATE_LIMIT_EVENTS.CONFIG_UPDATED, SendTarget.ALL_WINDOWS, {
+      providerId,
+      config: newConfig
+    })
+  }
+
+  private getProviderRateLimitConfig(providerId: string): RateLimitConfig {
+    const state = this.providerRateLimitStates.get(providerId)
+    return state?.config || this.DEFAULT_RATE_LIMIT_CONFIG
+  }
+
+  private canExecuteImmediately(providerId: string): boolean {
+    const state = this.providerRateLimitStates.get(providerId)
+    if (!state || !state.config.enabled) {
+      return true
+    }
+    const now = Date.now()
+    const intervalMs = (1 / state.config.qpsLimit) * 1000
+    return now - state.lastRequestTime >= intervalMs
+  }
+
+  private async executeWithRateLimit(providerId: string): Promise<void> {
+    const state = this.getOrCreateRateLimitState(providerId)
+    if (!state.config.enabled) {
+      this.recordRequest(providerId)
+      return Promise.resolve()
+    }
+    if (this.canExecuteImmediately(providerId)) {
+      this.recordRequest(providerId)
+      return Promise.resolve()
+    }
+    return new Promise<void>((resolve, reject) => {
+      const queueItem: QueueItem = {
+        id: `${providerId}-${Date.now()}-${Math.random()}`,
+        timestamp: Date.now(),
+        resolve,
+        reject
+      }
+
+      state.queue.push(queueItem)
+      console.log(
+        `[LLMProviderPresenter] Request queued for ${providerId}, queue length: ${state.queue.length}`
+      )
+      eventBus.send(RATE_LIMIT_EVENTS.REQUEST_QUEUED, SendTarget.ALL_WINDOWS, {
+        providerId,
+        queueLength: state.queue.length,
+        requestId: queueItem.id
+      })
+      this.processRateLimitQueue(providerId)
+    })
+  }
+
+  private recordRequest(providerId: string): void {
+    const state = this.getOrCreateRateLimitState(providerId)
+    const now = Date.now()
+    state.lastRequestTime = now
+    eventBus.send(RATE_LIMIT_EVENTS.REQUEST_EXECUTED, SendTarget.ALL_WINDOWS, {
+      providerId,
+      timestamp: now,
+      currentQps: this.getCurrentQps(providerId)
+    })
+  }
+
+  private async processRateLimitQueue(providerId: string): Promise<void> {
+    const state = this.providerRateLimitStates.get(providerId)
+    if (!state || state.isProcessing || state.queue.length === 0) {
+      return
+    }
+    state.isProcessing = true
+    try {
+      while (state.queue.length > 0) {
+        if (this.canExecuteImmediately(providerId)) {
+          const queueItem = state.queue.shift()
+          if (queueItem) {
+            this.recordRequest(providerId)
+            queueItem.resolve()
+            console.log(
+              `[LLMProviderPresenter] Request executed for ${providerId}, remaining queue: ${state.queue.length}`
+            )
+          }
+        } else {
+          const now = Date.now()
+          const intervalMs = (1 / state.config.qpsLimit) * 1000
+          const nextAllowedTime = state.lastRequestTime + intervalMs
+          const waitTime = Math.max(0, nextAllowedTime - now)
+          if (waitTime > 0) {
+            await new Promise((resolve) => setTimeout(resolve, waitTime))
+          }
+        }
+      }
+    } catch (error) {
+      console.error(
+        `[LLMProviderPresenter] Error processing rate limit queue for ${providerId}:`,
+        error
+      )
+      while (state.queue.length > 0) {
+        const queueItem = state.queue.shift()
+        if (queueItem) {
+          queueItem.reject(new Error('Rate limit processing failed'))
+        }
+      }
+    } finally {
+      state.isProcessing = false
+    }
+  }
+
+  private getOrCreateRateLimitState(providerId: string): ProviderRateLimitState {
+    let state = this.providerRateLimitStates.get(providerId)
+    if (!state) {
+      state = {
+        config: { ...this.DEFAULT_RATE_LIMIT_CONFIG },
+        queue: [],
+        lastRequestTime: 0,
+        isProcessing: false
+      }
+      this.providerRateLimitStates.set(providerId, state)
+    }
+    return state
+  }
+
+  private getCurrentQps(providerId: string): number {
+    const state = this.providerRateLimitStates.get(providerId)
+    if (!state || !state.config.enabled || state.lastRequestTime === 0) return 0
+    const now = Date.now()
+    const timeSinceLastRequest = now - state.lastRequestTime
+    const intervalMs = (1 / state.config.qpsLimit) * 1000
+    return timeSinceLastRequest < intervalMs ? 1 : 0
+  }
+
+  private getQueueLength(providerId: string): number {
+    const state = this.providerRateLimitStates.get(providerId)
+    return state?.queue.length || 0
+  }
+
+  private getLastRequestTime(providerId: string): number {
+    const state = this.providerRateLimitStates.get(providerId)
+    return state?.lastRequestTime || 0
+  }
+
+  private cleanupProviderRateLimit(providerId: string): void {
+    const state = this.providerRateLimitStates.get(providerId)
+    if (state) {
+      while (state.queue.length > 0) {
+        const queueItem = state.queue.shift()
+        if (queueItem) {
+          queueItem.reject(new Error('Provider removed'))
+        }
+      }
+      this.providerRateLimitStates.delete(providerId)
+      console.log(`[LLMProviderPresenter] Cleaned up rate limit state for ${providerId}`)
+    }
+  }
+
+  private onProvidersUpdated(providers: LLM_PROVIDER[]): void {
+    for (const provider of providers) {
+      if (provider.rateLimit) {
+        this.setProviderRateLimitConfig(provider.id, {
+          enabled: provider.rateLimit.enabled,
+          qpsLimit: provider.rateLimit.qpsLimit
+        })
+      }
+    }
+    const currentProviderIds = new Set(providers.map((p) => p.id))
+    const allStatus = this.getAllProviderRateLimitStatus()
+    for (const providerId of Object.keys(allStatus)) {
+      if (!currentProviderIds.has(providerId)) {
+        this.cleanupProviderRateLimit(providerId)
       }
     }
   }
